Add `-Z borrowck=migrate` flag, use it to link NLL up to AST-borrowck.
